add statically configured toBucket mapping and on-demand replications
It is now possible to add a bucket-mapping to the chorus storage
configuration. This allows replications, which are automatically
created, to use a custom destination bucket name.

Additionally, the possibility to only create replications for
source buckets included in the mapping was added. This makes it
possible to only create replications for a predefined subset of
available buckets in the main storage.

If 'onDemandReplication' is true in the storage configuration,
user replications will only be created if 'createReplications'
is set.

Bucketreplication are created on-the-fly if any writing request
is processed by the proxy.

Signed-off-by: Lukas Thiemeier <thiemeier@strato.de>