add transparent proxy mode
If the received request is already for the proxy's target domain, we can
forward it unaltered.

That saves computing a new signature for the request. It also protects from
errors from re-signing where the new signature is not computed correctly
(which is currently the case for v2 authentication when requests miss the
optional x-amz-content-sha256 header).

It is currently the only option to support "chunked uploads" requests
(https://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-streaming.html),
where the signature of each chunk is comprised of the previous chunk's
signature. That way, the first data chunk's signature depends on the
request's signature. If that is recomputed (which is required because it
depends on the host header), this means that all the other chunk signatures
will need to be recomputed as well, which is not implemented in chorus.

Signed-off-by: Jan Schlien <schlien@strato.de>