chorctl: adapt ui to new api
The chorus api does not return the size of the pending oan processed
objects anymore, which were used to calculate and display the progress
of a replication. The number of pending and processed objects is used
to display the progress now, which is as inaccurate as the size.

Instead of displaying the sizes, chorctl now displays the boolean
"IsInitDone". It is provided by the chorus-api and based on internal
queue statistics, which should be more accurate than the size and the
object count, as it consideres tha fact that objects might be processed
several times due to backend storate failures.

Signed-off-by: Lukas Thiemeier <thiemeier@strato.de>